pyproject.toml
[project]
name = "divoom-gaming-gate"
version = "0.1.11"
description = "A PyQt5-based character and screen manager for Divoom gaming devices."
authors = [{ name = "Your Name", email = "your@email.com" }]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "PyQt5",
    "PyQtWebEngine",
    "Pillow",
    "requests"
]
license = "GPL-3.0-or-later"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
divoom-gaming-gate = "divoom_gaming_gate.main:main"

[tool.setuptools.package-data]
"app" = ["GamingGate.ico"]
"divoom_gaming_gate.designer" = ["editor.html", "*.js", "*.css"]

[tool.briefcase]
project_name = "Divoom Gaming Gate"
bundle = "org.diastra.divoomgaminggate"

[tool.briefcase.app.divoom_gaming_gate]
formal_name = "Divoom Gaming Gate"
description = "A PyQt5-based character and screen manager for Divoom Times Gate."
icon = "divoom_gaming_gate/GamingGate"
sources = ["divoom_gaming_gate"]
entry_point = "divoom_gaming_gate.main:main"
license = "GPL-3.0-or-later"

